+++ src/helpers/environment.ts
@@ -1,0 +1,4 @@
+export const environment = {
+  production: 'https://api.parafin.com',
+  development: 'https://api.dev.parafin.com',
+}
